+++ frontend/src/components/Dash/Register.vue
@@ -40,11 +40,9 @@
 
 <script>
 import { useAccountStore } from "../../stores/account";
-import { useConfig } from "../../stores/config";
 export default {
 	data() {
 		return {
-			config: useConfig(),
 			errors: {
 				username: [],
 				password: [],