+++ frontend/src/components/Popups/ChangeEmail.vue
@@ -1,11 +1,10 @@
 <template>
 	<div class="changeEmail">
 		<h2>Change Email</h2>
-		<FormKit type="form" :actions="false" @submit="changeEmail">
+		<FormKit type="form" submit-label="Change Email" :submit-attrs="{ 'button-type': 'primary' }" @submit="changeEmail">
 			<FormKit type="email" label="New Email" v-model="newData.newEmail" autocomplete="email" />
 			<FormKit type="password" label="Password" v-model="newData.password" autocomplete="current-password" />
 			<FormKit type="text" label="Your 2FA token" v-model="newData.token" validation="number:required|length:6,6" v-if="account.account.totp"  autocomplete="one-time-code" />
-			<FormKit type="submit" label="Change Email" primary />
 			<p>{{ response }}</p>
 		</FormKit>
 	</div>