+++ frontend/src/components/Popups/ChangeUsername.vue
@@ -1,11 +1,10 @@
 <template>
 	<div class="changeUsername">
 		<h2>Change Username</h2>
-		<FormKit type="form" :actions="false" @submit="changeUsername">
+		<FormKit type="form" submit-label="Change Username" :submit-attrs="{ 'button-type': 'primary' }" @submit="changeUsername">
 			<FormKit type="text" label="New Username" v-model="newData.newUsername" autocomplete="username" />
 			<FormKit type="password" label="Password" v-model="newData.password" autocomplete="current-password" />
 			<FormKit type="text" label="Your 2FA token" v-model="newData.token" validation="number:required|length:6,6" v-if="account.account.totp"  autocomplete="one-time-code" />
-			<FormKit type="submit" label="Change Username" primary />
 			<p>{{ response }}</p>
 		</FormKit>
 	</div>