+++ src/api/middleware/requireLogin.js
@@ -1,6 +1,6 @@
 const { current: currentAccount, bySecret: getAccountBySecret } = require("../../db/modules/account/get");
 
-module.exports = (disallowSecret) => {
+module.exports = (disallowSecret, softFail) => {
 	return async (req, res, next) => {
 		let account;
 		let error;
@@ -9,24 +9,25 @@ module.exports = (disallowSecret) => {
 			if (disallowSecret) {
 				return res.status(403).json({
 					success: false,
-					message: "You cannot use this endpoint with your API secret"
-				})
+					message: "You cannot use this endpoint with your API secret",
+				});
 			}
 
 			[account, error] = await getAccountBySecret({
-				secret: req.body.secret || req.headers.authorization
+				secret: req.body.secret || req.headers.authorization,
 			});
 		} else {
 			[account, error] = await currentAccount(req);
 		}
 
-		if (error)
+		if (error && !softFail)
 			return res.status(error.code).json({
 				success: false,
 				message: error.message,
 			});
-	
-		req.account = account;
+
+		if (!error) req.account = account;
+
 		next();
-	}
+	};
 };