+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.3.9] - 2023-05-08
+
+- Properly handle logouts
+  - Logout would not properly work in local environment
+- Add about section to settings page
+- Properly use FormKit to disable form button while loading
+- Add github actions:
+  - Updating LICENSE year automatically
+  - Creating release on GitHub when PR merged
+
 ## [1.3.8] - 2023-05-04
 
 - Set image to lynx:1 in docker-compose example
@@ -59,4 +69,3 @@ First release, production ready
 - Main framework of Lynx
 - Import/Export
 - Add ShareX support
-